enhancements in test section.
 This involves:
 - Docmentation about fee workflow for every IT.
 - Refactor the IT's as almost IT's uses same code -> thus created some api's to reduce redundancy.
 - Fix some bugs in calculating coinswap fees.

+++ src/maker/api.rs
@@ -352,7 +352,7 @@ impl Maker {
                 )
                 .map_err(WalletError::Rpc)?
             {
-                if txout.confirmations < (REQUIRED_CONFIRMS as u32) {
+                if txout.confirmations < REQUIRED_CONFIRMS {
                     return Err(MakerError::General(
                         "funding tx not confirmed to required depth",
                     ));
